putting interface functions into emu2grab.py
primary mqtt work will be in emu2mqtt.py
